refactor tests for deprecation of coords class
