+++ tests/test_coordinates_new.py
@@ -2,15 +2,48 @@ import numpy as np
 import numpy.testing as npt
 import pytest
 
+import matplotlib.pyplot as plt
 from pyfar import Coordinates
-from pyfar.classes.coordinates import (sph2cart, cart2sph, cyl2cart)
+from pyfar.classes.coordinates import (sph2cart, cart2sph, cyl2cart, cart2cyl)
 from pyfar import (deg2rad, rad2deg)
+
+
+def test_coordinates_init():
+    """Test initialization of empty coordinates object."""
+    coords = Coordinates()
+    assert isinstance(coords, Coordinates)
 
 
 def test___eq___copy():
     coordinates = Coordinates(1, 2, 3, comment="Madre mia!")
     actual = coordinates.copy()
     assert coordinates == actual
+
+
+def test___eq___differInWeights_notEqual():
+    coordinates = Coordinates(1, 2, 3, weights=.5)
+    actual = Coordinates(1, 2, 3, weights=0.0)
+    assert not coordinates == actual
+
+
+def test___eq___differInShComment_notEqual():
+    coordinates = Coordinates(1, 2, 3, comment="Madre mia!")
+    actual = Coordinates(1, 2, 3, comment="Oh my woooooosh!")
+    assert not coordinates == actual
+
+
+@pytest.mark.parametrize(
+    'points_1, points_2, points_3, actual, expected', [
+        (1, 1, 1,                Coordinates(1, 1, -1),                 False),
+        ([1, 1], [1, 1], [1, 1], Coordinates([1, 1], [1, 1], [1, 2]),   False),
+        ([1, 1], [1, 1], [1, 1], Coordinates([1, 1.0], [1, 1.0], [1, 1]), True)
+    ])
+def test___eq___differInPoints(
+        points_1, points_2, points_3, actual, expected):
+    """This function checks against 3 different pairings of Coordinates."""
+    coordinates = Coordinates(points_1, points_2, points_3)
+    comparison = coordinates == actual
+    assert comparison == expected
 
 
 @pytest.mark.parametrize(
@@ -283,52 +316,16 @@ def test_find_slice_cart():
     d = np.linspace(-2, 2, 5)
 
     c = Coordinates(d, 0, 0)
-    index, mask = c.find_slice('x', 'met', 0, 1)
-    np.testing.assert_allclose(index[0], np.array([1, 2, 3]))
-    np.testing.assert_allclose(mask, np.array([0, 1, 1, 1, 0]))
+    mask = (c.x <= 1) & (c.x >= -1)
+    np.testing.assert_allclose(mask, np.array([0, 1, 1, 1, 0]) == 1)
 
     c = Coordinates(0, d, 0)
-    index, mask = c.find_slice('y', 'met', 0, 1)
-    np.testing.assert_allclose(index[0], np.array([1, 2, 3]))
-    np.testing.assert_allclose(mask, np.array([0, 1, 1, 1, 0]))
+    mask = (c.y <= 1) & (c.y >= -1)
+    np.testing.assert_allclose(mask, np.array([0, 1, 1, 1, 0]) == 1)
 
     c = Coordinates(0, 0, d)
-    index, mask = c.find_slice('z', 'met', 0, 1)
-    np.testing.assert_allclose(index[0], np.array([1, 2, 3]))
-    np.testing.assert_allclose(mask, np.array([0, 1, 1, 1, 0]))
-
-
-@pytest.mark.parametrize(
-    'coordinate, unit, value, tol, des_index, des_mask', [
-        ('azimuth', 'deg', 0, 1, np.array([1, 2, 3]),
-            np.array([0, 1, 1, 1, 0])),
-        ('azimuth', 'deg', 359, 2, np.array([0, 1, 2, 3]),
-            np.array([1, 1, 1, 1, 0])),
-        ('azimuth', 'deg', 1, 1, np.array([2, 3, 4]),
-            np.array([0, 0, 1, 1, 1])),
-    ])
-def test_find_slice_sph(coordinate, unit, value, tol, des_index, des_mask):
-    """Test different queries for find slice."""
-    # spherical grid
-    d = np.array([358, 359, 0, 1, 2]) * np.pi / 180
-    c = Coordinates.from_spherical_elevation(d, 0, 1)
-
-    index, mask = c.find_slice(coordinate, unit, value, tol)
-    np.testing.assert_allclose(index[0], des_index)
-    np.testing.assert_allclose(mask, des_mask)
-
-
-def test_find_slice_error():
-    d = np.array([358, 359, 0, 1, 2]) * np.pi / 180
-    c = Coordinates.from_spherical_elevation(d, 0, 1)
-    # out of range query
-    # with pytest.raises(AssertionError):
-    #     c.find_slice('azimuth', 'deg', -1, 1)
-    # non existing coordinate query
-    with pytest.raises(ValueError, match="does not exist"):
-        c.find_slice('elevation', 'ged', 1, 1)
-    with pytest.raises(ValueError, match="does not exist"):
-        c.find_slice('Ola', 'red', 1, 1)
+    mask = (c.z <= 1) & (c.z >= -1)
+    np.testing.assert_allclose(mask, np.array([0, 1, 1, 1, 0]) == 1)
 
 
 @pytest.mark.parametrize(
@@ -658,3 +655,191 @@ def test_angle_conversion_deg2rad():
     rad = deg2rad(deg)
     # check output values
     npt.assert_allclose(rad, np.atleast_2d([np.pi, 2*np.pi, 1]))
+
+
+def test_converters():
+    """
+    Test if converters can handle numbers (correctness of the conversion is
+    tested in test_setter_and_getter_with_conversion)
+    """
+    cart2sph(0, 0, 1)
+    sph2cart(0, 0, 1)
+    cart2cyl(0, 0, 1)
+    cyl2cart(0, 0, 1)
+
+
+def test_rotation_assertion():
+    """Test rotation with unknown rotation type."""
+    c = Coordinates(1, 0, 0)
+    # test with unknown type
+    with pytest.raises(ValueError):
+        c.rotate('urgh', 90)
+
+
+def test_inverse_rotation():
+    """Test the inverse rotation."""
+    xyz = np.concatenate((np.ones((2, 4, 1)),
+                          np.zeros((2, 4, 1)),
+                          np.zeros((2, 4, 1))), -1)
+    c = Coordinates(xyz[..., 0].copy(), xyz[..., 1].copy(), xyz[..., 2].copy())
+    c.rotate('z', 90)
+    c.rotate('z', 90, inverse=True)
+    npt.assert_allclose(c.cartesian, xyz, atol=1e-15)
+
+
+def test_coordinates_init_val():
+    """Test initializing Coordinates with values of different type and size."""
+
+    # test input: scalar
+    c1 = 1
+    # test input: 2 element vectors
+    c2 = [1, 2]                        # list
+    c3 = np.asarray(c2)                # flat np.array
+    # test input: 3 element vector
+    c6 = [1, 2, 3]
+    # test input: 2D matrix
+    c7 = np.array([[1, 2, 3], [1, 2, 3]])
+    # test input: 3D matrix
+    c8 = np.array([[[1, 2, 3], [1, 2, 3]],
+                   [[1, 2, 3], [1, 2, 3]]])
+
+    # tests that have to path
+    # input scalar coordinate
+    Coordinates(c1, c1, c1)
+    # input list of coordinates
+    Coordinates(c2, c2, c2)
+    # input scalar and lists
+    Coordinates(c1, c2, c2)
+    # input flat np.arrays
+    Coordinates(c3, c3, c3)
+    # input non flat vectors
+    # input 2D data
+    Coordinates(c1, c1, c7)
+    # input 3D data
+    Coordinates(c1, c1, c8)
+
+    # tests that have to fail
+    with pytest.raises(AssertionError):
+        Coordinates(c2, c2, c6)
+    with pytest.raises(AssertionError):
+        Coordinates(c6, c6, c7)
+    with pytest.raises(AssertionError):
+        Coordinates(c2, c2, c8)
+
+
+def test_coordinates_init_val_and_comment():
+    """Test initialization with comment."""
+    coords = Coordinates(1, 1, 1, comment='try this')
+    assert isinstance(coords, Coordinates)
+    assert coords.comment == 'try this'
+
+
+def test_coordinates_init_val_and_weights():
+    """Test initialization with weights."""
+    # correct number of weights
+    coords = Coordinates([1, 2], 0, 0, weights=[.5, .5])
+    assert isinstance(coords, Coordinates)
+    npt.assert_allclose(coords.weights, [.5, .5])
+
+    # incorrect number of weights
+    with pytest.raises(AssertionError):
+        Coordinates([1, 2], 0, 0, weights=.5)
+
+
+def test_show():
+    """Test if possible calls of show() pass."""
+    coords = Coordinates([-1, 0, 1], 0, 0)
+    # show without mask
+    coords.show()
+    # show with mask as list
+    coords.show([1, 0, 1])
+    # show with index as list
+    coords.show([0, 1])
+    # show with mask as ndarray
+    coords.show(np.array([1, 0, 1], dtype=bool))
+    # show with index as ndarray
+    coords.show(np.array([0, 1], dtype=int))
+
+    plt.close("all")
+
+
+def test_setter_weights():
+    """Test setting weights."""
+    coords = Coordinates([1, 2], 0, 0)
+    coords.weights = [.5, .5]
+    assert (coords.weights == np.array([.5, .5])).all()
+
+
+def test_setter_comment():
+    """Test setting the comment."""
+    coords = Coordinates()
+    coords.comment = 'now this'
+    assert coords.comment == 'now this'
+
+
+def test_cshape():
+    """Test the cshape attribute."""
+    # empty
+    coords = Coordinates()
+    assert coords.cshape == (0,)
+    # 2D points
+    coords = Coordinates([1, 0], [1, 1], [0, 1])
+    assert coords.cshape == (2,)
+    # 3D points
+    coords = Coordinates([[1, 2, 3], [4, 5, 6]], 1, 1)
+    assert coords.cshape == (2, 3)
+
+
+def test_cdim():
+    """Test the cdim attribute."""
+    # empty
+    coords = Coordinates()
+    assert coords.cdim == 0
+    # 2D points
+    coords = Coordinates([1, 0], 1, 1)
+    assert coords.cdim == 1
+    # 3D points
+    coords = Coordinates([[1, 2, 3], [4, 5, 6]], 1, 1)
+    assert coords.cdim == 2
+
+
+def test_csize():
+    """Test the csize attribute."""
+    # 0 points
+    coords = Coordinates()
+    assert coords.csize == 0
+    # two points
+    coords = Coordinates([1, 0], 1, 1)
+    assert coords.csize == 2
+    # 6 points in two dimensions
+    coords = Coordinates([[1, 2, 3], [4, 5, 6]], 1, 1)
+    assert coords.csize == 6
+
+
+def test_getitem():
+    """Test getitem with different parameters."""
+    # test without weights
+    coords = Coordinates([1, 2], 0, 0)
+    new = coords[0]
+    assert isinstance(new, Coordinates)
+    npt.assert_allclose(new.cartesian, np.atleast_2d([1, 0, 0]))
+
+    # test with weights
+    coords = Coordinates([1, 2], 0, 0, weights=[.1, .9])
+    new = coords[0]
+    assert isinstance(new, Coordinates)
+    npt.assert_allclose(new.cartesian, np.atleast_2d([1, 0, 0]))
+    assert new.weights == np.array(.1)
+
+    # test with 3D array
+    coords = Coordinates([[1, 2, 3, 4, 5], [2, 3, 4, 5, 6]], 0, 0)
+    new = coords[0:1]
+    assert isinstance(new, Coordinates)
+    assert new.cshape == (1, 5)
+
+    # test if sliced object stays untouched
+    coords = Coordinates([0, 1], [0, 1], [0, 1])
+    new = coords[0]
+    new.cartesian = np.array([2, 2, 2])
+    assert coords.cshape == (2,)
+    npt.assert_allclose(coords.cartesian[0], np.array([0, 0, 0]))
